Fix platform-specific errors on tvOS, Linux
